Frontend Rewrite 🚀
changes include

- responsive and sleek design
- added loader!!!
- updated templates
- used halfmoon.css for UI components
- improved performance
