--- expense app/website/static/sidenavbar_style.css
@@ -1,94 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    font-family: 'Roboto', sans-serif;
-    box-sizing: border-box;
-  }
-  
-  body {
-    background:linear-gradient(to right ,rgb(255, 255, 255),rgb(172, 239, 246));
-    
-  }
-  
-  .sidenav {
-    height: 100%;
-    width: 220px;
-    position: fixed;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    background-color: #086dac;
-    overflow-x: hidden;
-    padding-top: 0px;
-    color: #fff;
-    display: flex;
-    flex-direction: column;
-    border-right:5px solid #044065;
-  }
-
-  .sidenav a {
-    padding: 15px 10px;
-    text-decoration: none;
-    font-size: 25px;
-    color: #fff;
-    display: block;
-    transition: background-color 0.3s;
-  }
-
-  .sidenav a:hover {
-    background-color: #075c90;
-  }
-  
-  .sidenav a:last-child {
-    margin-top: auto;
-  }
-  
-  
-  .logo-container {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    justify-content: center;
-    margin-bottom: 15px;
-    background-color:#075c90;
-    height:25%;
-  }
-  
-  .logo-image {
-    max-width: 100%;
-    max-height: 100%;
-    width:auto;
-    height: auto; 
-}
-
-.logo-name {
-  font-size: 20px; 
-  color: #fff;
-  text-align: center;
-}
-  .content {
-    margin-left: 200px; 
-    padding: 20px;
-    color: #fff;
-  }
-  
-  .top-right {
-    position: absolute;
-    padding-bottom: 5px;
-    right: 10px;
-    margin-right: 10px;
-  }
-
-  .user-name {
-    font-size: 45px;
-    font-weight: bold;
-    color: white;
-  }
-
-  .user-icon {
-    font-size: 30px;
-    margin-right: 10px;
-    color: white;
-  }
-
-
