--- expense app/website/templates/add_expense.html
@@ -1,130 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Add Expense</title>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
-  <script src="https://kit.fontawesome.com/93782ebd13.js" crossorigin="anonymous"></script>
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='style.css') }}">
-  <script>
-    document.addEventListener("DOMContentLoaded", function() {
-      var flashMessages = document.querySelectorAll('.flash-message');
-  
-      setTimeout(function() {
-        hideFlashMessage(0);
-      }, 2000);
-  
-      function hideFlashMessage(index) {
-        if (index >= flashMessages.length) {
-          return;
-        }
-  
-        flashMessages[index].style.display = 'none';
-  
-        setTimeout(function() {
-          hideFlashMessage(index + 1);
-        }, 500);
-      }
-    });
-  </script>
-  
-</head>
-<body>
-  <div class="navbar">
-    <h1 class="navbar-title">Add Expense</h1>
-    {% with messages = get_flashed_messages(with_categories=true) %}
-      {% if messages %}
-        <ul class="flash-messages">
-          {% for category, message in messages %}
-            {% if category == 'error' %}
-              <li class="flash-message flash-error">{{ message }}</li>
-            {% elif category == 'success' %}
-              <li class="flash-message flash-success">{{ message }}</li>
-            {% endif %}
-          {% endfor %}
-        </ul>
-      {% endif %}
-    {% endwith %}
-    {% include 'side_navbar.html' %}
-  </div>  
-  <div class="container">
-    <div class="balance">
-      <h3>Current Balance: {{ total_balance }}</h3>
-    </div>
-    <!-- Expense Form -->
-    <form action="{{url_for('views.adding_expense')}}" method="POST">
-      <div class="input-group">
-        <label for="amount" class="input-group-text">Amount:</label>
-        <input type="number" id="amount" name="amount" required>
-        <span class="input-icon"><i class="fa-solid fa-indian-rupee-sign"></i></span>
-      </div>
-      <div class="input-group">
-        <div class="type-selection">
-          <div class="type-text">Type:</div>
-          <div class="radio-selection">
-            <label>
-              <input type="radio" name="amount_type" value="credit" required>
-              <span class="selection-text">Credit</span>
-            </label>
-            <label>
-              <input type="radio" name="amount_type" value="debit" required>
-              <span class="selection-text">Debit</span>
-            </label>
-          </div>
-        </div>
-      </div>
-      <div class="input-group">
-        <label for="description" class="input-group-text">Description:</label>
-        <textarea id="description" name="description" required></textarea>
-      </div>
-      <div class="input-group">
-        <label for="date" class="input-group-text">Date:</label>
-        <input type="date" id="date" name="date" required>
-      </div>
-      <div class="input-group">
-        <input type="submit" class="btn" value="Add Expense">
-      </div>
-    </form>
-    <!-- Expense Table -->
-    <h2>Expenses:</h2>
-    <table>
-      <tr>
-        <th>Amount</th>
-        <th>Description</th>
-        <th>Date</th>
-        <th>Type</th>
-        <th>Action</th>
-      </tr>
-      {% if expenses %}
-        {% for expense in expenses %}
-          <tr>
-            <td>
-              {% if expense.amount_type == "debit" %}
-                <span class="debit">-{{ expense.amount }}</span>
-              {% else %}
-                <span class="credit">{{ expense.amount }}</span>
-              {% endif %}
-            </td>
-            <td>{{ expense.description }}</td>
-            <td>{{ expense.date }}</td>
-            <td>
-              {% if expense.amount_type == "credit" %}
-                <span class="credit">Credit</span>
-              {% else %}
-                <span class="debit">Debit</span>
-              {% endif %}
-            </td>
-            <td>
-              <form action="{{ url_for('views.delete_expense', expense_id=expense.id) }}" method="POST" onsubmit="return confirm('Are you sure you want to delete this expense?');">
-                <input type="hidden" name="_method" value="DELETE">
-                <button type="submit" class="btn delete-button">
-                  <i class="fas fa-trash-alt"></i>
-                </button>
-              </form>
-            </td>
-          </tr>
-        {% endfor %}
-      {% endif %}
-    </table>
-  </div>
-</body>
-</html>
