+++ main.py
@@ -1,0 +1,6 @@
+from website import create_app
+
+app = create_app()
+
+if __name__=="__main__":
+    app.run(debug=False)
