+++ requirements.txt
@@ -1,0 +1,4 @@
+Flask==2.3.2
+Werkzeug==2.3.6
+Flask_Login==0.6.2
+flask_sqlalchemy==3.0.5
