+++ website/static/login_style.css
@@ -1,0 +1,205 @@
+* {
+    margin: 0;
+    padding: 0;
+    font-family: 'Roboto', sans-serif;
+    box-sizing: border-box;
+}
+body{
+    background: #073d5f;
+}
+.container {
+    width: 100%;
+    height: 100vh;
+    background-position: center;
+    background-size: cover;
+    position: relative;
+    animation: transitionIn 1.25s;
+}
+
+.form-box {
+    width: 90%;
+    max-width: 450px;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    border-radius: 10px;
+    transform: translate(-50%, -50%);
+    background: white;
+    padding: 50px 60px 50px;
+    text-align: center;
+    
+}
+
+.form-box h1 {
+    font-size: 30px;
+    margin-bottom: 60px;
+    color: #086dac;
+    position: relative;
+}
+
+.form-box h1::after {
+    content: '';
+    width: 30px;
+    height: 4px;
+    border-radius: 3px;
+    background: #086dac;
+    position: absolute;
+    bottom: -12px;
+    left: 50%;
+    transform: translateX(-50%);
+}
+
+.form-group {
+    background: #eaeaea;
+    margin: 15px 0;
+    border-radius: 5px;
+    display: flex;
+    align-items: center;
+    max-height: 65px;
+    transition: max-height 0.5s;
+    overflow: hidden;
+}
+
+input {
+    width: 100%;
+    background: transparent;
+    border: 0;
+    outline: 0;
+    padding: 18px 15px;
+}
+
+.form-group i {
+    margin-left: 15px;
+    color: #999;
+}
+
+form p {
+    text-align: left;
+    font-size: 13px;
+}
+
+.btn-field {
+    width: 100%;
+    margin: 0;
+    justify-content: center;
+    display: flex;
+    align-items: center;
+}
+
+.btn-field button {
+    flex-basis: 48%;
+    position: center;
+    background: #086dac;
+    color: #fff;
+    height: 40px;
+    border-radius: 10px;
+    border: 0;
+    outline: 0;
+    cursor: pointer;
+    transition: background 0.5s;
+}
+
+.btn-field button:hover {
+    background: #075c90;
+}
+
+.btn-field button:active {
+    background: #06456e;
+}
+
+.input-group {
+    height: 100%;
+    margin-bottom: 10px;
+}
+
+.btn-field button.disable {
+    background: #086dac;
+}
+
+.donthave {
+    padding: 3%;
+    font-size: 12px;
+    text-align: center;
+}
+
+.donthave p {
+    text-align: center;
+    font-size: 15px;
+}
+
+.donthave p a {
+    text-decoration: none;
+}
+
+
+.flash-container {
+    position: absolute;
+    top: 20px;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 100%;
+    max-width: 450px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+
+
+  
+  .flash-message {
+    margin: 5px 0;
+    padding: 10px;
+    border-radius: 5px;
+    justify-content: center;
+    color: #fff;
+    width: 250px;
+    text-align: center;
+  }
+  
+  .flash-success {
+    background-color: #28a745;
+  }
+  
+  .flash-error {
+    background-color: #dc3545;
+  }
+
+
+.flash-message {
+    opacity: 1;
+    transition: opacity 0.5s ease-out;
+}
+
+.flash-message.fade-out {
+    opacity: 0;
+}
+
+.logo-image {
+    max-width: 50%;
+    height: 50%;
+    border-radius: 100px;
+}
+
+.logo-container {
+    margin-bottom: 20px;
+    margin-bottom: 20px;
+    
+}
+
+#title {
+    margin-top: 20px;
+}
+
+
+  @keyframes transitionIn {
+    from{
+        transform: translateY(-10px);
+    }
+  
+    to{
+        transform: translateY(0);
+    }
+  }
+
+
